# math: use Sincos instead of Sin and Cos in Jn and Yn